+++ Src/Resource/Translations/RabbitCommon_nn.ts
@@ -931,28 +931,34 @@
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../RabbitCommonTools.cpp" line="430"/>
+        <location filename="../../RabbitCommonTools.cpp" line="431"/>
+        <location filename="../../RabbitCommonTools.cpp" line="441"/>
+        <source>Style</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../../RabbitCommonTools.cpp" line="451"/>
         <source>Log</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../RabbitCommonTools.cpp" line="434"/>
+        <location filename="../../RabbitCommonTools.cpp" line="455"/>
         <location filename="../../Log/DockDebugLog.cpp" line="82"/>
         <source>Open Log configure file</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../RabbitCommonTools.cpp" line="437"/>
+        <location filename="../../RabbitCommonTools.cpp" line="458"/>
         <source>Open Log file</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../RabbitCommonTools.cpp" line="440"/>
+        <location filename="../../RabbitCommonTools.cpp" line="461"/>
         <source>Open log folder</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../RabbitCommonTools.cpp" line="458"/>
+        <location filename="../../RabbitCommonTools.cpp" line="479"/>
         <source>Log dock</source>
         <translation type="unfinished"></translation>
     </message>